chore: extract bundle state type
