+++ crates/evm/execution-types/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "reth-execution-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+reth-primitives.workspace = true
+reth-trie.workspace = true
+reth-evm.workspace = true
+
+revm.workspace = true
+
+[features]
+optimism = []
