+++ crates/storage/provider/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 # reth
 reth-blockchain-tree-api.workspace = true
 reth-execution-errors.workspace = true
+reth-execution-types.workspace = true
 reth-primitives.workspace = true
 reth-fs-util.workspace = true
 reth-storage-errors.workspace = true
@@ -68,4 +69,4 @@ rand.workspace = true
 
 [features]
 test-utils = ["alloy-rlp", "reth-db/test-utils", "reth-nippy-jar/test-utils"]
-optimism = ["reth-primitives/optimism"]
+optimism = ["reth-primitives/optimism", "reth-execution-types/optimism"]
